feat(core): add junit5
